restructure: more similar to other packages
